+++ src/db.js
@@ -1,6 +1,5 @@
 import config from './config.json';
-// import freeradius_dictionary from './models/data/freeradius_dictionary';
-import sql from './models/data/freeradius_dictionary'
+import sql from './models/data/freeradius_dictionary';
 import Sequelize from 'sequelize';
 
 const Op = Sequelize.Op;
@@ -16,13 +15,13 @@ export default callback => {
   dbs.freeradius.radreply = dbs.freeradius.sequelize.import('./models/freeradius/radreply.js');
   dbs.freeradius.radcheck = dbs.freeradius.sequelize.import('./models/freeradius/radcheck.js');
   dbs.freeradius.radacct = dbs.freeradius.sequelize.import('./models/freeradius/radacct.js');
-  dbs.freeradius.dictionary = dbs.freeradius.sequelize.import('./models/freeradius/dictionary.js')
+  dbs.freeradius.dictionary = dbs.freeradius.sequelize.import('./models/freeradius/dictionary.js');
 
   dbs.freeradius.dictionary.count().then( attributes => {
     if (attributes === 0) {
-      sql.sqlRequests.forEach( sqlRequest => {
-        dbs.freeradius.sequelize.query(sqlRequest).then(results => {})
-      })
+      sql.sqlRequests.forEach( (sqlRequest) => {
+        dbs.freeradius.sequelize.query(sqlRequest).then( (results) => {});
+      });
     }
   });
 
@@ -49,6 +48,6 @@ export default callback => {
 	}).then(() => console.log('finished populating users'))
 		.catch(err => console.log('error populating users', err));
   
-  dbs.localDb.sequelize.sync().then(dbs.freeradius.sequelize.sync().then(callback(dbs)))
+  dbs.localDb.sequelize.sync().then(dbs.freeradius.sequelize.sync().then(callback(dbs)));
   
 }